OpenApi Client Generator (#1)
* POC for openApiGenerator

Trying to see if it's interesting to use an auto gen api

* Configures tsoa to dynamically set API host

Adds a pre-generation script to dynamically set the API host in the tsoa configuration based on the environment variable.
This allows the swagger documentation to accurately reflect the deployed environment.
Also updates nodemon to watch the pre-generation script for changes.

Removes the static tsoa.json and adds @types/node dependency.